+++ vue-hooks/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @zj-library/vue-hooks
 
+## 1.0.5
+
+### Patch Changes
+
+-   743532b: echarts setSeries函数dataKey类型必传的问题
+
+## 1.0.4
+
+### Patch Changes
+
+-   466974b: 修改类型不提示的问题
+-   Updated dependencies [466974b]
+    -   @zj-library/utils@1.0.6
+
 ## 1.0.3
 
 ### Patch Changes